+++ _examples/file-server/file-server/main.go
@@ -102,7 +102,7 @@ func uploadView(ctx iris.Context) {
 func upload(ctx iris.Context) {
 	ctx.SetMaxRequestBodySize(maxSize)
 
-	_, err := ctx.UploadFormFiles(uploadDir, beforeSave)
+	_, _, err := ctx.UploadFormFiles(uploadDir, beforeSave)
 	if err != nil {
 		ctx.StopWithError(iris.StatusPayloadTooRage, err)
 		return
@@ -111,12 +111,13 @@ func upload(ctx iris.Context) {
 	ctx.Redirect("/files")
 }
 
-func beforeSave(ctx iris.Context, file *multipart.FileHeader) {
+func beforeSave(ctx iris.Context, file *multipart.FileHeader) bool {
 	ip := ctx.RemoteAddr()
 	ip = strings.ReplaceAll(ip, ".", "_")
 	ip = strings.ReplaceAll(ip, ":", "_")
 
 	file.Filename = ip + "-" + file.Filename
+	return true
 }
 
 func deleteFile(ctx iris.Context) {